+++ src/app/login/page.tsx
@@ -4,47 +4,33 @@ import { Card, CardHeader, CardTitle, CardDescription, CardAction, CardFooter, C
 import { Input } from "~/components/ui/input"
 import { Label } from '~/components/ui/label';
 import { Button } from '~/components/ui/button';
+import SignupPage from '../signup/page';
+import Link from 'next/link';
 
-export default function SignupPage() {
+export default function LoginPage() {
     return(
-        <div className='min-h-screen flex items-center justify-center'>
+        <div className='min-h-screen flex items-center justify-center bg-[#FFF1D4]'>
             <div className='w-1/4'>
                 <Card className="w-full">
                 <CardHeader>
-                    <CardTitle>Sign up</CardTitle>
+                    <CardTitle>Welcome Back</CardTitle>
                 </CardHeader>
                 <CardContent>
                     <form>
                     <div className="flex flex-col gap-3">
                         <div className="grid gap-1">
                         <Input
-                            id="username"
+                            id="usernameEmail"
                             type="text"
-                            placeholder="Username"
+                            placeholder="Username or Email"
                             required
                         />
                         </div>
                         <div className="grid gap-1">
                         <Input
-                            id="email"
-                            type="email"
-                            placeholder="Email"
-                            required
-                        />
-                        </div>
-                        <div className="grid gap-2">
-                        <Input
                             id="password"
                             type="password"
                             placeholder="Password"
-                            required
-                        />
-                        </div>
-                        <div className="grid gap-2">
-                        <Input
-                            id="passwordConfirmation"
-                            type="password"
-                            placeholder="Confirm Password"
                             required
                         />
                         </div>
@@ -53,13 +39,15 @@ export default function SignupPage() {
                 </CardContent>
                 <CardFooter className="flex-col gap-2">
                     <Button variant="brownPrimary" type="submit" className="w-full">
-                    Create Account
+                    Log in
                     </Button>
                 </CardFooter>
                 </Card>
                 
                 <div className='flex justify-center text-yellow-600 mt-6'>
-                    <p className='text-2xl font-normal'>Not a newbie around here? <span className='font-bold cursor-pointer'>Log in</span></p>
+                    <p className='text-2xl font-normal'>New around here? <span className='font-bold cursor-pointer'>
+                        <Link href='/signup'>Sign up</Link>
+                    </span></p>
                 </div>
             </div>
         </div>
