+++ src/app/page.tsx
@@ -2,8 +2,7 @@ import Link from "next/link";
 
 import { LatestPost } from "~/app/_components/post";
 import { api, HydrateClient } from "~/trpc/server";
-import SignupPage from "./_components/signup/page";
-import LoginPage from "./_components/login/page";
+import LoginPage from "./login/page";
 
 export default async function Home() {
   const hello = await api.post.hello({ text: "from tRPC" });
@@ -12,7 +11,6 @@ export default async function Home() {
 
   return (
     <div className="min-h-screen bg-[#FFF1D4]">
-      {/* <SignupPage /> */}
       <LoginPage />
       
       {/* <main className="flex min-h-screen flex-col items-center justify-center bg-gradient-to-b from-[#2e026d] to-[#15162c] text-white"> */}
